requirements.txt
torch>=2.0
numpy
matplotlib
Pillow
